fix(seeders): centralize FRED proxy helpers in _seed-utils; apply to all stlouisfed callers
Moves resolveProxy/curlFetch/fredFetchJson to _seed-utils.mjs.
Applies proxy routing to all 4 seeders that call api.stlouisfed.org:
- seed-economy.mjs (FRED series + JPY fallback)
- seed-supply-chain-trade.mjs (shipping indices)
- seed-bls-series.mjs (BLS labor series)
- seed-economic-calendar.mjs (release dates)
