+++ scripts/seed-economy.mjs
@@ -1,40 +1,10 @@
 #!/usr/bin/env node
 
-import { loadEnvFile, CHROME_UA, runSeed, writeExtraKeyWithMeta, sleep } from './_seed-utils.mjs';
-import { execFileSync } from 'child_process';
+import { loadEnvFile, CHROME_UA, runSeed, writeExtraKeyWithMeta, sleep, resolveProxy, fredFetchJson } from './_seed-utils.mjs';
 
 loadEnvFile(import.meta.url);
 
-// Proxy for FRED — Railway container IPs get rate-limited/blocked by api.stlouisfed.org.
-// Supports PROXY_URL="host:port:user:pass" (Decodo) or OREF_PROXY_AUTH="user:pass@host:port" (Froxy).
-function resolveProxy() {
-  const raw = process.env.PROXY_URL || '';
-  if (raw) {
-    const parts = raw.split(':');
-    if (parts.length === 4) {
-      const [host, port, user, pass] = parts;
-      return `${user}:${pass}@${host.replace(/^gate\./, 'us.')}:${port}`;
-    }
-    return raw;
-  }
-  return process.env.OREF_PROXY_AUTH || '';
-}
 const _proxyAuth = resolveProxy();
-
-// curl-based fetch for sources that block Railway IPs.
-// Returns response body as string; throws on non-2xx.
-function curlFetch(url, headers = {}) {
-  const args = ['-sS', '--compressed', '--max-time', '15', '-L'];
-  if (_proxyAuth) args.push('-x', `http://${_proxyAuth}`);
-  for (const [k, v] of Object.entries(headers)) args.push('-H', `${k}: ${v}`);
-  args.push('-w', '\n%{http_code}');
-  args.push(url);
-  const raw = execFileSync('curl', args, { encoding: 'utf8', timeout: 20000, stdio: ['pipe', 'pipe', 'pipe'] });
-  const nl = raw.lastIndexOf('\n');
-  const status = parseInt(raw.slice(nl + 1).trim(), 10);
-  if (status < 200 || status >= 300) throw Object.assign(new Error(`HTTP ${status}`), { status });
-  return raw.slice(0, nl);
-}
 
 // ─── Keys (must match handler cache keys exactly) ───
 const KEYS = {
@@ -191,16 +161,9 @@ async function fetchFredSeries() {
         series_id: seriesId, api_key: apiKey, file_type: 'json',
       });
 
-      const fredFetchJson = async (url) => {
-        if (_proxyAuth) return JSON.parse(curlFetch(url, { Accept: 'application/json' }));
-        const r = await fetch(url, { headers: { Accept: 'application/json' }, signal: AbortSignal.timeout(10_000) });
-        if (!r.ok) throw Object.assign(new Error(`HTTP ${r.status}`), { status: r.status });
-        return r.json();
-      };
-
       const [obsResp, metaResp] = await Promise.allSettled([
-        fredFetchJson(`https://api.stlouisfed.org/fred/series/observations?${obsParams}`),
-        fredFetchJson(`https://api.stlouisfed.org/fred/series?${metaParams}`),
+        fredFetchJson(`https://api.stlouisfed.org/fred/series/observations?${obsParams}`, _proxyAuth),
+        fredFetchJson(`https://api.stlouisfed.org/fred/series?${metaParams}`, _proxyAuth),
       ]);
 
       if (obsResp.status === 'rejected') {
@@ -288,7 +251,7 @@ async function fetchFredJpyFallback() {
   if (!apiKey) return [];
   try {
     const params = new URLSearchParams({ series_id: 'DEXJPUS', api_key: apiKey, file_type: 'json', sort_order: 'desc', limit: '250' });
-    const data = await fetchJsonSafe(`https://api.stlouisfed.org/fred/series/observations?${params}`, 10_000);
+    const data = await fredFetchJson(`https://api.stlouisfed.org/fred/series/observations?${params}`, _proxyAuth);
     return (data.observations || [])
       .map((o) => { const v = parseFloat(o.value); return Number.isNaN(v) || o.value === '.' ? null : v; })
       .filter(Boolean)
